+++ generator/src/main/java/com/scottlogic/deg/generator/profile/constraints/atomic/NotMatchesStandardConstraint.java
@@ -18,7 +18,7 @@ package com.scottlogic.deg.generator.profile.constraints.atomic;
 import com.scottlogic.deg.common.profile.Field;
 import com.scottlogic.deg.generator.fieldspecs.FieldSpec;
 import com.scottlogic.deg.generator.fieldspecs.FieldSpecFactory;
-import com.scottlogic.deg.generator.restrictions.StringRestrictionsFactory;
+import com.scottlogic.deg.generator.restrictions.string.StringRestrictionsFactory;
 
 import java.util.Objects;
 import java.util.regex.Pattern;
